pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pandakinetics"
version = "0.1.0"
description = "Multi-Scale Structure-Kinetics Simulator for Drug Design"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "PandaKinetics Team", email = "contact@pandakinetics.org"}
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "matplotlib>=3.5.0",
    "networkx>=2.6",
    "tqdm>=4.60.0",
    "loguru>=0.6.0",
    "click>=8.0.0",
    "pydantic>=1.8.0",
    "scikit-learn>=1.0.0",
]

[project.optional-dependencies]
ml = [
    "torch>=1.12.0",
    "torchvision>=0.13.0",
]
gpu = [
    "cupy-cuda11x>=10.0.0; sys_platform=='linux'",
]
chem = [
    "rdkit>=2022.3.1",
    "biotite>=0.36.0",
    "mdtraj>=1.9.6",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
full = ["pandakinetics[ml,gpu,chem]"]

[project.scripts]
pandakinetics = "pandakinetics.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pandakinetics*"]
exclude = ["tests*"]
